+++ 01-starting-project/src/components/CoreConcept.jsx
@@ -1,0 +1,20 @@
+// Normally we pass and use props with a single parameter
+// function CoreConcept(props) {
+// But we can also destructure the data passed from REST operator & props and this will work with both.
+
+export default function CoreConcept({
+  title: TITLE,
+  image: IMAGE,
+  description: DESCRIPTION,
+}) {
+  const altText = "No image found";
+  return (
+    <>
+      <li>
+        <img src={IMAGE} alt={altText} />
+        <h3>{TITLE}</h3>
+        <p>{DESCRIPTION}</p>
+      </li>
+    </>
+  );
+}
